🏗️ Refactor components and update imports
- Removed CardListNoImage component and its associated code.
- Updated CardListWithImage component to include a limit prop for item display.
- Changed import paths for AppLogo, AppLogoIcon, and Icon components to reflect new structure.
- Deleted unused components: AsidePageOrganism, ListJournalPage, WithImageLandscapeOrganism, and ContainerCardsWithImageLandscape.
- Moved CardListNoImage and CardListWithImage to the molecules directory.
- Refactored date formatting in various pages to use formatDate utility.
- Updated nav-footer and app-header to use new icon imports.
- Removed deprecated components: HeadingSmall, Heading, InputError, TextLink, and Icon.
