+++ resources/js/components/atoms/app-logo-icon.tsx
