+++ resources/js/components/atoms/heading-small.tsx
