+++ resources/js/components/atoms/heading.tsx
