+++ resources/js/components/atoms/input-error.tsx
