--- resources/js/components/containers/container-cards-with-image-landscape.tsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import CardWithImageLandscape from '@/components/atoms/card-with-image-landscape'
-
-const ContainerCardsWithImageLandscape = () => {
-  return (
-    <div>
-        
-    </div>
-  )
-}
-
-export default ContainerCardsWithImageLandscape
