+++ resources/js/components/molecules/CardListNoImage.tsx
