--- resources/js/components/organisms/aside-page-organism.tsx
@@ -1,9 +1,0 @@
-const AsidePageOrganism = () => {
-    return (
-        <div className="">
-            <h2>AsidePage</h2>
-        </div>
-    );
-};
-
-export default AsidePageOrganism;
