--- resources/js/components/organisms/list-journal-page.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const ListJournalPage = () => {
-  return (
-    <div>list-journal-page</div>
-  )
-}
-
-export default ListJournalPage
