--- resources/js/components/organisms/withImage-landscape-organism.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const WithImageLandscapeOrganism = () => {
-  return (
-    <div className="">
-      <h2>WithImageLandscapeOrganism</h2>
-    </div>
-  )
-}
-
-export default WithImageLandscapeOrganism
